# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "selfevals",
    "description": "HTTP bridge for the selfevals evals framework.",
    "version": "0.0.1"
  },
  "paths": {
...
    "/api/workspaces/{workspace_id}/failure-modes": {
      "get": {
        "tags": [
          "failure-modes"
        ],
        "summary": "Failure Modes Index",
        "operationId": "failure_modes_index_api_workspaces__workspace_id__failure_modes_get",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by status (candidate/official/retired).",
              "title": "Status"
            },
            "description": "Filter by status (candidate/official/retired)."
          },
          {
            "name": "X-SelfEvals-User",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Caller identity. In `local`/`header` mode: a plain user id. In `token` mode: a signed token from `selfevals.api.tokens.issue_token`.",
              "title": "X-Selfevals-User"
            },
            "description": "Caller identity. In `local`/`header` mode: a plain user id. In `token` mode: a signed token from `selfevals.api.tokens.issue_token`."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailureModeListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspaces/{workspace_id}/failure-modes/{failure_mode_id}/promote": {
      "post": {
        "tags": [
          "failure-modes"
        ],
        "summary": "Failure Modes Promote",
        "operationId": "failure_modes_promote_api_workspaces__workspace_id__failure_modes__failure_mode_id__promote_post",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          },
          {
            "name": "failure_mode_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Failure Mode Id"
            }
          },
          {
            "name": "X-SelfEvals-User",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Caller identity. In `local`/`header` mode: a plain user id. In `token` mode: a signed token from `selfevals.api.tokens.issue_token`.",
              "title": "X-Selfevals-User"
            },
            "description": "Caller identity. In `local`/`header` mode: a plain user id. In `token` mode: a signed token from `selfevals.api.tokens.issue_token`."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailureModeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspaces/{workspace_id}/failure-modes/{failure_mode_id}/retire": {
      "post": {
        "tags": [
          "failure-modes"
        ],
        "summary": "Failure Modes Retire",
        "operationId": "failure_modes_retire_api_workspaces__workspace_id__failure_modes__failure_mode_id__retire_post",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          },
          {
            "name": "failure_mode_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Failure Mode Id"
            }
          },
          {
            "name": "X-SelfEvals-User",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Caller identity. In `local`/`header` mode: a plain user id. In `token` mode: a signed token from `selfevals.api.tokens.issue_token`.",
              "title": "X-Selfevals-User"
            },
            "description": "Caller identity. In `local`/`header` mode: a plain user id. In `token` mode: a signed token from `selfevals.api.tokens.issue_token`."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailureModeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspaces/{workspace_id}/failure-modes/{failure_mode_id}/merge": {
      "post": {
        "tags": [
          "failure-modes"
        ],
        "summary": "Failure Modes Merge",
        "operationId": "failure_modes_merge_api_workspaces__workspace_id__failure_modes__failure_mode_id__merge_post",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          },
          {
            "name": "failure_mode_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Failure Mode Id"
            }
          },
          {
            "name": "X-SelfEvals-User",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Caller identity. In `local`/`header` mode: a plain user id. In `token` mode: a signed token from `selfevals.api.tokens.issue_token`.",
              "title": "X-Selfevals-User"
            },
            "description": "Caller identity. In `local`/`header` mode: a plain user id. In `token` mode: a signed token from `selfevals.api.tokens.issue_token`."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MergeFailureModeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailureModeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspaces/{workspace_id}/failure-modes/{failure_mode_id}": {
      "patch": {
        "tags": [
          "failure-modes"
        ],
        "summary": "Failure Modes Edit",
        "operationId": "failure_modes_edit_api_workspaces__workspace_id__failure_modes__failure_mode_id__patch",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          },
          {
            "name": "failure_mode_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Failure Mode Id"
            }
          },
          {
            "name": "X-SelfEvals-User",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Caller identity. In `local`/`header` mode: a plain user id. In `token` mode: a signed token from `selfevals.api.tokens.issue_token`.",
              "title": "X-Selfevals-User"
            },
            "description": "Caller identity. In `local`/`header` mode: a plain user id. In `token` mode: a signed token from `selfevals.api.tokens.issue_token`."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditFailureModeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailureModeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EditFailureModeRequest": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "definition": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Definition"
          }
        },
        "type": "object",
        "title": "EditFailureModeRequest",
        "description": "Patch a mode's human-facing text. At least one field must be set."
      },
...
      "FailureModeListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/FailureModeResponse"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "title": "FailureModeListResponse"
      },
...
      "FailureModeResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "definition": {
            "type": "string",
            "title": "Definition"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "parent_mode_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Mode Id"
          },
          "proposed_by": {
            "type": "string",
            "title": "Proposed By"
          },
          "example_count": {
            "type": "integer",
            "title": "Example Count",
            "default": 0
          },
          "first_seen_iteration": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Seen Iteration"
          },
          "superseded_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Superseded By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "slug",
          "title",
          "definition",
          "status",
          "proposed_by",
          "created_at",
          "updated_at"
        ],
        "title": "FailureModeResponse",
        "description": "One failure mode, projected for the taxonomy UI."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MergeFailureModeRequest": {
        "properties": {
          "into_id": {
            "type": "string",
            "minLength": 1,
            "title": "Into Id"
          }
        },
        "type": "object",
        "required": [
          "into_id"
        ],
        "title": "MergeFailureModeRequest",
        "description": "Merge this mode's examples into `into_id`, then retire the source."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}